server/requirements.txt
fastapi
uvicorn
motor
pydantic
fastapi-jwt-auth
authlib
pytonapi
python-dotenv
bcrypt
pyjwt
python-telegram-bot